BlockRacing SUmmary
Block-racing is a car racing game based on Ethereum blockchain, a gaming platform that gamers play and bet money on, it is also a venture that we are inviting Bitcoin/Ethereum geeks like you to invest in. 

As a game controlled by smart contracts, BlockRacing gaming proceeds are divided and distributed to shareholders accounts automatically. The equity will soon be traded in exchanges of crypto currencies worldwide.

As an investment, BlockRacing can be either a short term or long term investment. You can trade BlockRacing share with anyone in a P2P fashion, without intermediaries, as you do with Bitcoin or any digital currencies. You can sell your BlockRacing share for Bitcoin, ETH or US dollars anytime. 

In the long run, with increasing size of the fund pool, proceeds and profit to game players and shareholders will be on the rise as well. 

We are working hard on development and will roll out the game soon. We have launched an ICO (Initial Coin Offering) for BlockRacing. For d